Load image comments action request image comments from the loader.
I went straight to modular approach with MVP design with `Adapter` additional layer.
But to not over-deliver the solution already I weakified `ImageCommentsPresenter` view prototype properties to adress that "something" is not good with this approach and to justify next changes.
For now there would be plain retain cycles, that are showing when testing this code